moved col_download_unread into journalist_app/utils.py
