+++ securedrop/journalist.py
@@ -23,7 +23,8 @@ from journalist_app.utils import (get_source, commit_account_changes,
                                   make_password, set_diceware_password,
                                   make_star_true, make_star_false, download,
                                   delete_collection, confirm_bulk_delete,
-                                  bulk_delete, col_download_all)
+                                  bulk_delete, col_download_all,
+                                  col_download_unread)
 
 app = create_app(config)
 
@@ -436,22 +437,6 @@ def col_process():
 
     method = actions[action]
     return method(cols_selected)
-
-
-def col_download_unread(cols_selected):
-    """Download all unread submissions from all selected sources."""
-    submissions = []
-    for filesystem_id in cols_selected:
-        id = Source.query.filter(Source.filesystem_id == filesystem_id) \
-                   .one().id
-        submissions += Submission.query.filter(
-            Submission.downloaded == false(),
-            Submission.source_id == id).all()
-    if submissions == []:
-        flash(gettext("No unread submissions in selected collections."),
-              "error")
-        return redirect(url_for('index'))
-    return download("unread", submissions)
 
 
 def col_star(cols_selected):
